fix: main 함수를 간략화 & 모든 역할이 Application 클래스에 있던 것을 분산
utils 폴더에 game을 진행하면서 필요한 함수들과 숫자관련 처리를 위해 필요한 함수들을 구조화함
코드를 좀 더 확장성 있게 변경하였으며 추후 유지보수를 위해 역할이 비슷한 것들을 한 곳에 모아둠
