+++ src/main/java/baseball/utils/GameStatus.java
@@ -1,0 +1,5 @@
+package baseball.utils;
+
+public enum GameStatus {
+    RESTART, EXIT;
+}
